requirements.txt
# Monte Carlo 验证实验依赖
# 安装命令: pip install -r requirements.txt

torch>=2.0.0
numpy>=1.21.0
matplotlib>=3.5.0
pyyaml>=6.0
pytest>=7.0.0
